refactor : used page
기존 best page에서 사용하던 Category component refactoring
Used component 생성 및 리팩토링

Related to : #44
