+++ src/app/best/page.tsx
@@ -1,9 +1,9 @@
-import Category from '@/components/best/category/Category';
-import Rank from '@/components/best/rank/Rank';
 import PageHeader from '@/components/common/PageHeader';
 import Sort from '@/components/common/Sort';
 import styles from '@/styles/best/best.module.css';
 import Pagination from '@/components/common/Pagination';
+import Category from '@/components/common/Category';
+import Rank from '@/components/best/Rank';
 
 export default function bestPage() {
 	return (
