+++ src/app/used/page.tsx
@@ -1,8 +1,8 @@
-import Category from '@/components/best/category/Category';
-import BookItem from '@/components/common/BookItem';
+import Category from '@/components/common/Category';
 import Pageheader from '@/components/common/PageHeader';
 import Pagination from '@/components/common/Pagination';
 import Sort from '@/components/common/Sort';
+import Used from '@/components/used/Used';
 import styles from '@/styles/used/used.module.css';
 
 export default function usedPage() {
@@ -13,18 +13,7 @@ export default function usedPage() {
 				<div className={styles.wrapper}>
 					<Category />
 					<Sort />
-					<div className={styles.usedWrapper}>
-						<BookItem />
-						<BookItem />
-						<BookItem />
-						<BookItem />
-						<BookItem />
-						<BookItem />
-						<BookItem />
-						<BookItem />
-						<BookItem />
-						<BookItem />
-					</div>
+					<Used />
 					<Pagination />
 				</div>
 			</div>
